Move polynomial search functions into `_search.py`
